requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
Django==3.2.3
Flask==2.0.0
idna==2.10
image==1.5.33
itsdangerous==2.0.0
Jinja2==3.0.0
lxml==4.6.3
MarkupSafe==2.0.0
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
pkg-resources==0.0.0
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.4
Werkzeug==2.0.0
